requirements.txt
Django>=4.2,<5.0
django-ninja>=1.0.0
pyjwt>=2.8.0